fix: Refactoring resolvable importer
