+++ backend/data/src/main/kotlin/io/tolgee/service/key/resolvableImport/ResolvableKeyImporterContext.kt
@@ -1,0 +1,9 @@
+package io.tolgee.service.key.resolvableImport
+
+import io.tolgee.model.key.Key
+import io.tolgee.service.key.resolvableImport.ResolvableImporter.TranslationToModify
+
+class ResolvableKeyImporterContext(val key: Key, val rootContext: ResolvableImportContext) {
+  val wasPlural = key.isPlural
+  val translationsToModify = mutableListOf<TranslationToModify>()
+}
